Re-enable automatic PATH adjustment due to what seems like a bug in TeXLive 2020
TeXLive installation can be controlled via environment variables.
These are set in the Dockerfile, and have been working.
However, when the requested TeXLive version is 'latest',
TeXLive 2020 did not install into the 'latest' path,
but into the '2020' path. Symlinking only the 'latest' path
into /usr/local/bin (on PATH), this broke installation (the
2020 was not on PATH).

I don't know what's wrong, but if we allow TeXLive to modify the
PATH on its own, it will do so successfully, no matter the
installation directory we requested and did or did not get.

This only works for modern TeXLive releases; older ones have to
have their PATH adjusted manually. Therefore, still do this.