Provide USER for container
This helps security but also convenience, see the comments in the
Dockerfile.

The new user will have UID/GID 1000/1000, which will have the
generated PDFs be owned by a normal user. If the host user running
Docker is also 1000/1000 (good chance for single-user Linux machines,
as far as I can see), file ownership will be solved and correct.